training/requirements_train.txt
tensorflow==2.17.0
numpy==1.26.4
Pillow==10.4.0
matplotlib==3.9.0
scikit-learn==1.5.0
seaborn==0.13.2
